refactor: remove unnecessary new helper files
